stores IDs in 64-bit integer [clang] [FORTRAN]
COMMENTS:
Stores the particle IDs in a signed 64-bit integer for
interoperability with FORTRAN. Note that the size (number of bytes)
used for the size_t depends on the implementation so my only
option was to use an integer of 64-bits since all the fields of
the sphere_t are 64-bits by design (and I want to keep it that
way for simplicity).

This change is necessary for it is unusual to use doubles for
addressing arrays as will be done with the list of neighbors.
I was aware of this from the beginning but decided to do this
change just before working on the neighbor lists.

the code passes the existing tests

valgrind reports no memory issues

+++ src/sphere.c
@@ -84,7 +84,7 @@ sphere_t* create ()
   double* t_y = spheres -> t_y;
   double* t_z = spheres -> t_z;
   double* list = spheres -> list;
-  double* id = spheres -> id;
+  int64_t* id = spheres -> id;
 
   zeros(size_x, x);
   zeros(size_y, y);
