+++ src/util.h
@@ -6,7 +6,7 @@
 
 bool sign (double const x);
 void zeros (size_t const size, double* x);
-void iota (size_t const size, double* x);
+void iota (size_t const size, int64_t* x);
 
 void mask_partition (size_t const size, const double* restrict x, double* restrict mask);
 void mask_unlimited (size_t const size,
